+++ cmd/cryptd/main.go
@@ -10,13 +10,17 @@ import (
 func main() {
 	if len(os.Args) < 2 {
 		fmt.Fprintln(os.Stderr, "usage: cryptd <command> [args]")
-		fmt.Fprintln(os.Stderr, "commands: serve, validate")
+		fmt.Fprintln(os.Stderr, "commands: serve, headless, autoplay, validate")
 		os.Exit(1)
 	}
 
 	switch os.Args[1] {
 	case "serve":
 		runServe(os.Args[2:])
+	case "headless":
+		runHeadless(os.Args[2:])
+	case "autoplay":
+		runAutoplay(os.Args[2:])
 	case "validate":
 		runValidate(os.Args[2:])
 	default: